flask/requirements.txt
# flask
flask

# dependencies
requests
slackclient
slackeventsapi

python-dotenv

# rasa
rasa_core

# Database
redis
